Split tmux input encoding and hotkey tests
